test: refactor tests to reflect new structure
